refactor : Refactoring Package racingcar
refactor
 - Change package name racingcar -> domain
 - Create package named view
 - Move Classes InputOutput and Validator into package view
