+++ src/main/java/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package domain;
 
 import java.util.ArrayList;
