+++ src/main/java/view/Validator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package view;
 
 import java.util.Arrays;
 import java.util.HashSet;
