+++ src/test/java/domain/ValidatorTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package domain;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import view.Validator;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
